Methods for MID
Implemented the new and improved methods for MID. Including two methods: SD and MVC, both with two settings. The two new threshold methods for the MVC, must be selectable on the GUI - Not yet.
The switch cases for the SD methods, are inaprropriately named, but should work: SD05 = SD*3 and SD03 = SD*2. The proposed implementation for SD, is based on the old framework, updating the mean per 33 samples, and only after the grand average of 3 seconds of data has been recorded, will the SD be calculated (based on the proper mean).
The MVC method currently assumes a first run of the switch case 'st_th', which should result in THRESHOLD = MEAN. The proposed idea is then to run a second switch case: 'st_MVC', during the MVC recording. This switch case is very similar to 'st_th' and results in a THRESHOLD, dependent on which version of the MVC method you choose.  Whether this new switch case of 'st_MVC' is the best way to include the MVC part of the threshold calculation, I do not know, but it seems to me as the simplest solution.